requirements.txt
Flask==2.2.5
requests==2.25.1
python-dotenv==0.19.0
python-jose==3.3.0
